Use caption node and support cross-referencing.
This changes the tikz node to be enumerable, so if the numfig option is
enabled in Sphinx then the tikz pictures are numbered along with
standard figures.

To look the same as a standard figure caption, a child
docutils.nodes.caption node is added rather than storing the string.
When visiting the tikz node, no longer raise SkipNode so Sphinx renders
the caption node for us. We then need a depart() method to finish the
block.

Finally, to support cross-referencing with the :numref: role, the LaTeX
output needs a label added to each figure environment. Sphinx provides
the functions to do this, so we just need to make a call in the depart()
method.