Use \capstart so hyperlinks jump to the correct point.
By default, hyperlinks to a figure will jump to where the label is
defined, which is after the caption. This means the figure is usually
off-screen. Sphinx uses the hypcap package to fix this; it requires
placing \capstart where the link should jump to. This commit places it
after \centering which is the same as the standard Sphinx figure.