readme: add texlive-latex-extra dependency
While using `sphinxcontrib.tikz`, I encountered an error similar to
that documented in
https://github.com/JuliaTeX/TikzPictures.jl/issues/68, namely, that
`standalone.cls` was not found. On Ubuntu Linux 22.04 (and probably
earlier), this file is installed with the `texlive-latex-extra`
package.

To help other users avoid this error, this commit adds that package to
the list of Ubuntu Linux package dependencies for
`sphinxcontrib.tikz` discussed in `README.rst`.

Installing this package also ensures that the `amsmath` LaTeX packages
are installed. Those LaTeX packages are part of the
`texlive-latex-base` package, which is a dependency of
`texlive-latex-extra`.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>